LayoutTests/platform/mac-ventura/imported/w3c/web-platform-tests/mathml/relations/css-styling/width-height-001-expected.txt


FAIL width and height properties on maction assert_approx_equals: width expected 500 +/- 1 but got 0
FAIL inline-size and block-size properties on maction assert_approx_equals: width expected 600 +/- 1 but got 0
FAIL width and height properties on menclose assert_approx_equals: width expected 500 +/- 1 but got 11.15625
FAIL inline-size and block-size properties on menclose assert_approx_equals: width expected 600 +/- 1 but got 11.15625
FAIL width and height properties on merror assert_approx_equals: width expected 500 +/- 1 but got 0
FAIL inline-size and block-size properties on merror assert_approx_equals: width expected 600 +/- 1 but got 0
FAIL width and height properties on mfrac assert_approx_equals: width expected 500 +/- 1 but got 0
FAIL inline-size and block-size properties on mfrac assert_approx_equals: width expected 600 +/- 1 but got 0
FAIL width and height properties on mi assert_approx_equals: width expected 500 +/- 1 but got 501.59375
FAIL inline-size and block-size properties on mi assert_approx_equals: width expected 600 +/- 1 but got 601.59375
FAIL width and height properties on mmultiscripts assert_approx_equals: width expected 500 +/- 1 but got 0.625
FAIL inline-size and block-size properties on mmultiscripts assert_approx_equals: width expected 600 +/- 1 but got 0.625
PASS width and height properties on mn
PASS inline-size and block-size properties on mn
PASS width and height properties on mo
PASS inline-size and block-size properties on mo
FAIL width and height properties on mover assert_approx_equals: width expected 500 +/- 1 but got 0
FAIL inline-size and block-size properties on mover assert_approx_equals: width expected 600 +/- 1 but got 0
FAIL width and height properties on mpadded assert_approx_equals: width expected 500 +/- 1 but got 0
FAIL inline-size and block-size properties on mpadded assert_approx_equals: width expected 600 +/- 1 but got 0
FAIL width and height properties on mphantom assert_approx_equals: width expected 500 +/- 1 but got 0
FAIL inline-size and block-size properties on mphantom assert_approx_equals: width expected 600 +/- 1 but got 0
FAIL width and height properties on mroot assert_approx_equals: width expected 500 +/- 1 but got 8.375
FAIL inline-size and block-size properties on mroot assert_approx_equals: width expected 600 +/- 1 but got 8.375
FAIL width and height properties on mrow assert_approx_equals: width expected 500 +/- 1 but got 0
FAIL inline-size and block-size properties on mrow assert_approx_equals: width expected 600 +/- 1 but got 0
PASS width and height properties on ms
PASS inline-size and block-size properties on ms
FAIL width and height properties on mspace assert_approx_equals: width expected 500 +/- 1 but got 0
FAIL inline-size and block-size properties on mspace assert_approx_equals: width expected 600 +/- 1 but got 0
FAIL width and height properties on msqrt assert_approx_equals: width expected 500 +/- 1 but got 12.703125
FAIL inline-size and block-size properties on msqrt assert_approx_equals: width expected 600 +/- 1 but got 12.703125
FAIL width and height properties on mstyle assert_approx_equals: width expected 500 +/- 1 but got 0
FAIL inline-size and block-size properties on mstyle assert_approx_equals: width expected 600 +/- 1 but got 0
FAIL width and height properties on msub assert_approx_equals: width expected 500 +/- 1 but got 0.625
FAIL inline-size and block-size properties on msub assert_approx_equals: width expected 600 +/- 1 but got 0.625
FAIL width and height properties on msubsup assert_approx_equals: width expected 500 +/- 1 but got 0.625
FAIL inline-size and block-size properties on msubsup assert_approx_equals: width expected 600 +/- 1 but got 0.625
FAIL width and height properties on msup assert_approx_equals: width expected 500 +/- 1 but got 0.625
FAIL inline-size and block-size properties on msup assert_approx_equals: width expected 600 +/- 1 but got 0.625
PASS width and height properties on mtable
PASS inline-size and block-size properties on mtable
PASS width and height properties on mtext
PASS inline-size and block-size properties on mtext
FAIL width and height properties on munder assert_approx_equals: width expected 500 +/- 1 but got 0
FAIL inline-size and block-size properties on munder assert_approx_equals: width expected 600 +/- 1 but got 0
FAIL width and height properties on munderover assert_approx_equals: width expected 500 +/- 1 but got 0
FAIL inline-size and block-size properties on munderover assert_approx_equals: width expected 600 +/- 1 but got 0
FAIL width and height properties on semantics assert_approx_equals: width expected 500 +/- 1 but got 0
FAIL inline-size and block-size properties on semantics assert_approx_equals: width expected 600 +/- 1 but got 0
